refactor(layout): decompose artifact placement search
